Marginal Statistics

This is the main function of the package. It calculates

average observed,

average predicted,

partial dependence, and

counts/weights over (possibly binned) features X specified by their column names v.

marginal(object, ...)

# Default S3 method
marginal(
  object,
  v,
  data,
  y = NULL,
  pred = NULL,
  pred_fun = stats::predict,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  wprob_low = 0.01,
  wprob_high = 0.99,
  calc_pred = TRUE,
  pd_n = 500L,
  ...
)

# S3 method for class 'ranger'
marginal(
  object,
  v,
  data,
  y = NULL,
  pred = NULL,
  pred_fun = NULL,
  w = NULL,
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  wprob_low = 0.01,
  wprob_high = 0.99,
  calc_pred = TRUE,
  pd_n = 500L,
  ...
)

# S3 method for class 'explainer'
marginal(
  object,
  v,
  data = object[["data"]],
  y = NULL,
  pred = NULL,
  pred_fun = object[["predict_function"]],
  w = object[["weights"]],
  breaks = "Sturges",
  right = TRUE,
  discrete_m = 2L,
  wprob_low = 0.01,
  wprob_high = 0.99,
  calc_pred = TRUE,
  pd_n = 500L,
  ...
)

Arguments

object

Fitted model.

...

Further arguments passed to pred_fun(), e.g., type = "response" in a glm() or (typically) prob = TRUE in binary probabilistic models.

v

Vector of variable names to calculate statistics.

data

Matrix or data.frame.

y

Numeric vector with observed values of the response. Can also be a column name in data. Omitted if NULL (default).

pred

Numeric vector with predictions. If NULL, it is calculated as pred_fun(object, data, ...). Used to save time if marginal() is to be called multiple times.

pred_fun

Prediction function, by default stats::predict. The function takes three arguments (names irrelevant): object, data, and ....

w

Optional vector with case weights. Can also be a column name in data.

breaks

An integer, a vector, a string or a function specifying the bins of the numeric X variables, and passed to graphics::hist(). The default is "Sturges". Not vectorized over v.

right

Should bins created via graphics::hist() be right-closed? The default is TRUE. Vectorized over v. Only relevant for numeric X.

discrete_m

Numeric X variables with up to this number of unique values should not be binned. The default is 2. Vectorized over v.

wprob_low

Small values of numeric X variables are capped at this quantile. Set to 0 to avoid Winsorizing. Note that at most 100k observations are sampled to calculate the quantiles (uses random seed). Vectorized over v.

wprob_high

High values of numeric X variables are capped at this quantile. Set to 1 to avoid Winsorizing. Note that at most 100k observations are sampled to calculate the quantile (uses random seed). Vectorized over v.

calc_pred

Should predictions be calculated? Default is TRUE. Only relevant if pred = NULL.

pd_n

Size of the data used for calculating partial dependence. The default is 500. For larger data (and w), pd_n rows are randomly sampled (uses random seed). Each variable specified by v uses the same subsample. Set to 0 to omit.

Value

A list (of class "marginal") with a data.frame of statistics per feature. Use single bracket subsetting to select part of the output.

Methods (by class)

marginal(default): Default method.

marginal(ranger): Method for "ranger" models.

marginal(explainer): Method for DALEX "explainer".

See also

Examples

fit <- lm(Sepal.Length ~ ., data = iris)
xvars <- colnames(iris)[-1]
M <- marginal(fit, v = xvars, data = iris, y = "Sepal.Length", breaks = 5)
#> Error in win_prob(x, probs = c(wprob_low, wprob_high), nmax = 1e+05): could not find function "win_prob"
M
#> Error: object 'M' not found
M |> plot()
#> Error: object 'M' not found