Plots Marginal Object

Plots marginal statistics using a color blind palette from "ggthemes". You can switch to plotly by setting backend = "plotly".

# S3 method for class 'marginal'
plot(
  x,
  ncol = 2L,
  byrow = TRUE,
  share_y = FALSE,
  ylim = NULL,
  cat_lines = TRUE,
  num_points = FALSE,
  title = "",
  subplot_titles = TRUE,
  ylab = NULL,
  line_names = c(obs = "y_mean", pred = "pred_mean", pd = "pd"),
  colors = getOption("marginalplot.colors"),
  fill = getOption("marginalplot.fill"),
  bar_height = 1,
  bar_width = 0.7,
  bar_measure = c("weight", "N"),
  wrap_x = 10,
  rotate_x = 0,
  backend = getOption("marginalplot.backend"),
  ...
)

Arguments

x

An object of class "marginal".

ncol

Number of columns in the plot layout. Only for multiple plots.

byrow

Should plots be placed by row? Default is TRUE. Only for multiple plots.

share_y

Should y axis be shared across all subplots? No effect if ylim is passed. Only for multiple plots.

ylim

A vector of length 2 with manual y axis limits.

cat_lines

Show lines for non-numeric features. Default is TRUE.

num_points

Show points for numeric features. Default is FALSE.

title

Overall plot title, by default "" (no title).

subplot_titles

Should variable names be shown as subplot titles? Default is TRUE. Only for multiple plots.

ylab

Label of the y axis. The default NULL automatically derives a reasonable name based on the calculated statistics.

line_names

Named vector controlling the legend labels, the lines shown, and the line order. By default c(obs = "y_mean", pred = "pred_mean", pd = "pd"). The names of the vector will be shown as legend labels. The values refer to the calculated statistics.

colors

Line colors corresponding to (available) line_names. By default, a color blind friendly palette from "ggthemes", namely c("#CC79A7", "#009E73", "#56B4E9"). To change globally, set options(marginalplot.colors = new colors).

fill

Fill color of bars. The default equals "lightgrey". To change globally, set options(marginalplot.fill = new color).

bar_height

Relative bar height (default 1). Set to 0 for no bars.

bar_width

Relative bar width of non-numeric features, by default 0.7.

bar_measure

What should bars represent? Either "weight" (default) or "N".

wrap_x

Should categorical x axis labels be wrapped after this length? The default is 10. Set to 0 for no wrapping. Vectorized over x. Only for "ggplot2" backend.

rotate_x

Should categorical xaxis labels be rotated by this angle? The default is 0 (no rotation). Vectorized over x. Only for "ggplot2" backend.

backend

Plot backend, either "ggplot2" (default) or "plotly". To change globally, set options(marginalplot.backend = "plotly").

...

Passed to patchwork::plot_layout() or plotly::subplot(). Typically not used.

Value

If single plot, an object of class "ggplot" or "plotly". Otherwise, an object of class "patchwork", or a "plotly" subplot.

See also