Plots "multimarginal" Object

Plots "multimarginal" Object

# S3 method for class 'multimarginal'
plot(
  x,
  ncols = 2L,
  share_y = FALSE,
  sort = FALSE,
  ylim = NULL,
  scale_exposure = 1,
  line_colors = getOption("marginalplot.line_colors"),
  fill = getOption("marginalplot.fill"),
  wrap_x = 10,
  rotate_x = 0,
  backend = getOption("marginalplot.backend"),
  ...
)

Arguments

x

An object of class "marginal".

ncols

Number of columns in the plot layout.

share_y

Should y axis be shared across all subplots? No effect if ylim is passed.

sort

Should plots be sorted in decreasing order of main effect importance measured by the exposure weighted variance of partial dependence? Only if partial dependence values have been calculated (i.e., not available for average_observed()).

ylim

Manual y axis range.

scale_exposure

Scale of the exposure bars, by default 1. Set to 0 for no bars. Currently, values between 0 and 1 are possible only with backend "ggplot2".

line_colors

Named vector of line colors. By default, a color blind palette from "ggthemes" is used, equaling to c(obs = "#E69F00", pred = "#009E73", pd = "#56B4E9"). To change globally, set options(marginalplot.line_colors = "named color vector"). Can be used to remove certain lines in the plot.

fill

Fill color of bars. The default equals "lightgrey". To change globally, set options(marginalplot.fill = "new color").

wrap_x

Should categorical x axis labels be wrapped after this number of characters? The default is 10. Only for "ggplot2" backend.

rotate_x

Should categorical x axis labels be rotated by this number of degrees? Only for "ggplot2" backend. The default is 0 (no rotation).

backend

Plotting backend, either "ggplot2" (default) or "plotly". To change globally, set options(marginalplot.backend = "plotly").

...

Currently not used.

Value

Object of class "ggplot" (single plot) or "patchwork" (multiple plots), or "plotly"/"htmlwidget" (with plotly backend).

See also