Plots "multimarginal" Object

Plots "multimarginal" Object

# S3 method for class 'multimarginal'
plot(
  x,
  line_colors = getOption("marginalplot.line_colors"),
  fill = getOption("marginalplot.fill"),
  ncols = 2L,
  share_y = FALSE,
  ylim = NULL,
  show_exposure = TRUE,
  wrap_x = 10,
  rotate_x = 0,
  backend = getOption("marginalplot.backend"),
  ...
)

Arguments

x

An object of class "marginal".

line_colors

Named vector of line colors. By default, a color blind palette from "ggthemes" is used, equalling to c(obs = "#E69F00", pred = "#009E73", pd = "#56B4E9"). To change globally, set options(marginalplot.line_colors = "named color vector"). Can be used to remove certain lines in the plot.

fill

Fill color of bars. The default equals "lightgrey". To change globally, set options(marginalplot.fill = "new color").

ncols

Number of columns in the plot layout.

share_y

Should y axis be shared across all subplots? This has no effect if ylim is provided.

ylim

Manual limits of the y axis range.

show_exposure

Should exposure bars be shown (on hidden right y axis)? For backend = "ggplot", a value between 0 and 1 scales the bar heights so they would cover only part of the y range.

wrap_x

Should categorical x axis labels be wrapped after this number of characters? The default is 10. Only "with ggplot2" backend.

rotate_x

Should categorical x axis labels be rotated by this number of degrees? Only "with ggplot2" backend. The default is 0 (no rotation).

backend

Plotting backend, one of "ggplot2" (default) or "plotly". To change globally, set options(marginalplot.backend = "plotly").

...

Internally used.